// packages/datadog-api-client-v2/models/UserAttributes.d.ts
/**
 * Unless explicitly stated otherwise all files in this repository are licensed under the Apache-2.0 License.
 * This product includes software developed at Datadog (https://www.datadoghq.com/).
 * Copyright 2020-Present Datadog, Inc.
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
/**
* Attributes of user object returned by the API.
*/
export declare class UserAttributes {
    /**
    * Creation time of the user.
    */
    'createdAt'?: Date;
    /**
    * Whether the user is disabled.
    */
    'disabled'?: boolean;
    /**
    * Email of the user.
    */
    'email'?: string;
    /**
    * Handle of the user.
    */
    'handle'?: string;
    /**
    * URL of the user's icon.
    */
    'icon'?: string;
    /**
    * Time that the user was last modified.
    */
    'modifiedAt'?: Date;
    /**
    * Name of the user.
    */
    'name'?: string;
    /**
    * Status of the user.
    */
    'status'?: string;
    /**
    * Title of the user.
    */
    'title'?: string;
    /**
    * Whether the user is verified.
    */
    'verified'?: boolean;
    static readonly discriminator: string | undefined;
    static readonly attributeTypeMap: {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static getAttributeTypeMap(): {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static deserialize(data: {
        [key: string]: any;
    }): UserAttributes;
    static serialize(data: UserAttributes): {
        [key: string]: any;
    };
    constructor();
}