// packages/datadog-api-client-v2/models/UserResponseRelationships.d.ts
/**
 * Unless explicitly stated otherwise all files in this repository are licensed under the Apache-2.0 License.
 * This product includes software developed at Datadog (https://www.datadoghq.com/).
 * Copyright 2020-Present Datadog, Inc.
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RelationshipToOrganization } from './RelationshipToOrganization';
import { RelationshipToOrganizations } from './RelationshipToOrganizations';
import { RelationshipToRoles } from './RelationshipToRoles';
import { RelationshipToUsers } from './RelationshipToUsers';
/**
* Relationships of the user object returned by the API.
*/
export declare class UserResponseRelationships {
    'org'?: RelationshipToOrganization;
    'otherOrgs'?: RelationshipToOrganizations;
    'otherUsers'?: RelationshipToUsers;
    'roles'?: RelationshipToRoles;
    static readonly discriminator: string | undefined;
    static readonly attributeTypeMap: {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static getAttributeTypeMap(): {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static deserialize(data: {
        [key: string]: any;
    }): UserResponseRelationships;
    static serialize(data: UserResponseRelationships): {
        [key: string]: any;
    };
    constructor();
}